♻️ separate controller wire validation
