+++ libs/contracts/src/agent-controller-skill-workload.validator.ts
@@ -1,7 +1,7 @@
 import { z } from "zod";
 
 import type { AgentControllerSkillWorkloadAssignmentCommand, AgentControllerSkillWorkloadAssignmentResult, AgentControllerSkillWorkloadClaim, AgentControllerSkillWorkloadPodRegistrationCommand, AgentControllerSkillWorkloadPodRegistrationResult, AgentControllerSkillWorkloadReleaseClaim, AgentControllerSkillWorkloadReleaseCommand, AgentControllerSkillWorkloadReleaseResult } from "./agent-controller-skill-workload.types.js";
-import { _AgentControllerBoundedIdentifierSchema, _AgentControllerMillisecondInstantSchema, _AgentControllerPositiveIntegerSchema, _ParseAgentControllerCommand, _ParseAgentControllerModel } from "./agent-controller.validator.js";
+import { _AgentControllerBoundedIdentifierSchema, _AgentControllerMillisecondInstantSchema, _AgentControllerPositiveIntegerSchema, _ParseAgentControllerCommand, _ParseAgentControllerModel } from "./agent-controller-wire.validator.js";
 
 /**
  * Zod validators live beside the governed skill workload wire models so controller and server
