+++ libs/contracts/src/index.ts
@@ -62,7 +62,8 @@ export { type ManagedRunInputScopeAttachment, type RunInputSnapshot, type RunInp
 export { type AgentControllerSkillWorkloadAssignmentCommand, type AgentControllerSkillWorkloadAssignmentResult, type AgentControllerSkillWorkloadClaim, type AgentControllerSkillWorkloadPodRegistrationCommand, type AgentControllerSkillWorkloadPodRegistrationResult, type AgentControllerSkillWorkloadReleaseClaim, type AgentControllerSkillWorkloadReleaseCommand, type AgentControllerSkillWorkloadReleaseResult } from "./agent-controller-skill-workload.types.js";
 export { ___ParseAgentControllerSkillWorkloadAssignmentCommand, ___ParseAgentControllerSkillWorkloadAssignmentResult, ___ParseAgentControllerSkillWorkloadClaim, ___ParseAgentControllerSkillWorkloadPodRegistrationCommand, ___ParseAgentControllerSkillWorkloadPodRegistrationResult, ___ParseAgentControllerSkillWorkloadReleaseClaim, ___ParseAgentControllerSkillWorkloadReleaseCommand, ___ParseAgentControllerSkillWorkloadReleaseResult } from "./agent-controller-skill-workload.validator.js";
 export { AGENT_CONTROLLER_PROJECTED_TOKEN_AUDIENCE, AGENT_CONTROLLER_SERVICE_ACCOUNT_NAME, type AgentControllerRunAttemptAssignmentCommand, type AgentControllerRunAttemptAssignmentResult, type AgentControllerRunAttemptClaim, type AgentControllerRunAttemptClaimLease, type AgentControllerRunAttemptProjection, type AgentControllerRunOutboxPruneResult, type AgentControllerRunWorkloadRegistrationCommand, type AgentControllerRunWorkloadRegistrationResult, type AgentControllerRunWorkloadReleaseClaim, type AgentControllerRunWorkloadReleaseProjection } from "./agent-controller.types.js";
-export { ___IsAgentControllerIdentifier, ___IsEmptyAgentControllerCommand, ___ParseAgentControllerOutboxPrunedCount, ___ParseAgentControllerRunAttemptAssignmentCommand, ___ParseAgentControllerRunAttemptAssignmentResult, ___ParseAgentControllerRunAttemptClaim, ___ParseAgentControllerRunWorkloadRegistrationCommand, ___ParseAgentControllerRunWorkloadRegistrationResult, ___ParseAgentControllerRunWorkloadReleaseClaim } from "./agent-controller.validator.js";
+export { ___IsAgentControllerIdentifier, ___IsEmptyAgentControllerCommand } from "./agent-controller-wire.validator.js";
+export { ___ParseAgentControllerOutboxPrunedCount, ___ParseAgentControllerRunAttemptAssignmentCommand, ___ParseAgentControllerRunAttemptAssignmentResult, ___ParseAgentControllerRunAttemptClaim, ___ParseAgentControllerRunWorkloadRegistrationCommand, ___ParseAgentControllerRunWorkloadRegistrationResult, ___ParseAgentControllerRunWorkloadReleaseClaim } from "./agent-controller.validator.js";
 export { AGENT_RUNTIME_PROJECTED_TOKEN_AUDIENCE, AGENT_RUNTIME_PROTOCOL_V1, MANAGED_AGENT_RUNTIME_PROFILE_NAME, MANAGED_AGENT_RUNTIME_PROJECTED_TOKEN_AUDIENCE, ___IsAgentRuntimeServiceAccountName, ___IsManagedAgentRuntimeServiceAccountName, type AgentRuntimeProjectedTokenAudience, type AgentRuntimeProtocolVersion, type ManagedAgentRuntimeProfileName, type ManagedAgentRuntimeProjectedTokenAudience, type CancelAttemptCommand, type ResumeAttemptCommand, type RuntimeCandidate, type RuntimeCandidateCoordinates, type RuntimeCommand, type RuntimeCommandCoordinates, type RuntimeCommandEnvelope, type RuntimeEventCandidate, type RuntimeExternalActionCandidate, type RuntimeStreamOpen, type StartAttemptCommand } from "./agent-runtime-protocol.types.js";
 export { type RuntimeAssignment, type RuntimeAssignmentIdentity, type ServiceRuntimeAssignmentIdentity, type UserRuntimeAssignmentIdentity } from "./runtime-assignment.types.js";
 export { type TenantModelSet } from "./tenant-models.types.js";
